Consolidate clob-wire's derives and shrink its header
A macro_rules! wrapper states the anchor/wincode/IDL derive set once,
instead of repeating the same six attribute lines on every type. The
crate header drops from a 31-line //! block to a short // block; it
stays // rather than becoming a crate doc, because clob-wire-v2 pulls
this file in with include!, and include! cannot splice an inner doc
comment into a module that has already started parsing.

CrankAccountV0::writable is now bool instead of a 0/nonzero u8; every
producer already wrote only 0 or 1. FILL_BATCH_CEILING moves here from
the CLOB program, next to ORDER_VIEW_CEILING, and the CLOB re-exports
it.

CrankResolverV0::program, CrankAccountV0::address and
OrderRulesV0::place_authority stay [u8; 32] rather than becoming
Address. Anchor's IDL build derives IdlBuild for its own Pubkey
re-export, not for a bare solana_address::Address, so velocity's own
`bun run program:idl` fails to compile the moment one of these fields
is an Address. The doc comments say so at each field.

CrankResolverV0 restates relay_spec::CrankSpecV0 and CrankAccountV0
restates relay_spec::AccountRefV0, because relay-spec depends on
nothing but bytemuck by design and derives neither wincode's schema
pair nor anchor's borsh pair.

+++ anchor-v2/programs/clob/src/instructions/set_crank_conditions_v0.rs
@@ -47,7 +47,7 @@ pub fn handle_set_crank_conditions_v0(
         .iter()
         .map(|account| AccountRefV0 {
             address: account.address,
-            writable: account.writable,
+            writable: account.writable as u8,
         })
         .collect();
     let watched = ctx.accounts.market.address().to_bytes();
